pyproject.toml
[project]
name = "amazon-listing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.3.0",
    "litellm>=1.72.6.post1",
    "python-dotenv>=1.1.0",
]
